# src/shared/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wallet Service API",
    "version": "1.0.0",
    "description": "Wallet service with Paystack deposits, JWT (Google) auth and API key access."
  },
  "servers": [{ "url": "/" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "detail": { "type": "object" }
        }
      },
...
      "ApiKeyCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "wallet-service" },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["deposit", "transfer", "read"]
            },
            "example": ["deposit", "read"]
          },
          "expiry": {
            "type": "string",
            "description": "1H,1D,1M,1Y",
            "example": "1D"
          }
        },
        "required": ["name", "permissions", "expiry"]
      },
      "ApiKeyResponse": {
        "type": "object",
        "properties": {
          "api_key": { "type": "string", "example": "sk_live_xxxxx" },
          "name": { "type": "string", "example": "wallet-service" },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T12:00:00Z"
          }
        }
      },
      "RolloverRequest": {
        "type": "object",
        "properties": {
          "expired_key_id": {
            "type": "string",
            "example": "FGH2485K6KK79GKG9GKGK"
          },
          "expiry": { "type": "string", "example": "1M" }
        },
        "required": ["expired_key_id", "expiry"]
      },
...
    }
  },
  "paths": {
...
    "/keys/create": {
      "post": {
        "tags": ["Keys"],
        "summary": "Create API key",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ApiKeyCreate" },
              "examples": {
                "create": {
                  "value": {
                    "name": "wallet-service",
                    "permissions": ["deposit", "read", "transfer"],
                    "expiry": "1D"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyResponse" },
                "examples": {
                  "created": {
                    "value": {
                      "api_key": "sk_live_xxxxx",
                      "expires_at": "2025-01-01T12:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request / limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/keys/rollover": {
      "post": {
        "tags": ["Keys"],
        "summary": "Rollover an expired API key",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RolloverRequest" },
              "examples": {
                "roll": {
                  "value": {
                    "expired_key_id": "FGH2485K6KK79GKG9GKGK",
                    "expiry": "1M"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New API key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyResponse" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}